Uniformisation des animations (`default` depuis --animation-*, `spring`)
