+++ src/layout/header/scrolling.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 import posed from "react-pose";
 import {PoseElementProps} from "react-pose/lib/components/PoseElement/types";
 
+import {springPose} from "../../animation";
 import {ScrollableContext} from "../../components";
 import {useTheme} from "../../theme";
 
@@ -49,11 +50,11 @@ export function HeaderScrolling({canDeploy, children, theme: pTheme}: HeaderScro
 const Header = posed.header({
     enter: {
         y: "0%",
-        transition: {type: "spring", stiffness: 170, damping: 26}
+        ...springPose
     },
     exit: {
         y: "-105%",
-        transition: {type: "spring", stiffness: 170, damping: 26}
+        ...springPose
     }
 });
